fix issues found while testing
